RQ-58-SPLIT (#242) unit 1: relocate the pattern-match/fallback path out of the monolith
Pure relocation, verified verbatim: `select()`, `apply_replacement`,
`f32_trunc_range_guard`, and `select_default` (2,503 contiguous lines) move
from instruction_selector.rs into instruction_selector/select_default.rs as
a second `impl InstructionSelector` block. The ONLY token change inside the
moved block is `fn select_default` -> `pub(super) fn select_default`, needed
because its one cross-layer caller (`select_with_stack`'s `_ =>`
fallthrough) now lives in the parent module.

Organizing principle (the #197 seam): the parent keeps the shared lowering
core (operand/stack machinery, bounds-check emission, VFP lowering, call
marshalling, config) and the production direct path `select_with_stack`
(forced by --relocatable — direct != default, verified against
arm_backend.rs:605 before moving anything); the sibling holds the legacy
pattern-match path, production-unreachable for covered ops.

Byte identity: 10/10 frozen anchors green, full workspace suite green —
zero emitted bytes moved (a pure move cannot move bytes; the gate confirms).

Ledger (same commit, family glob from b37994da makes relocation read ±0):
- selector_lines_code 17912 -> 17938: +26 module plumbing only (sibling doc
  header, `use super::*;`, impl wrapper, parent `mod` decl) — waived.
- selector_lines_total 28533 -> 28559 (track).
- mirror_marker_files 58 -> 59: three existing "mirror" comment lines moved
  into a second file; obligation population unchanged — waived.
- Wildcard pins, sel_dsl_rules, mirror_obligation_files: unmoved.

Refs #242
