+++ artifacts/status.json
@@ -17,7 +17,7 @@
   "isa_model_simplified_only": 68,
   "isa_model_uncovered": 4,
   "kani_harnesses": 18,
-  "mirror_marker_files": 58,
+  "mirror_marker_files": 59,
   "mirror_obligation_files": 23,
   "rocq_admit_tactics": 2,
   "rocq_admitted": 2,
@@ -28,8 +28,8 @@
   "sel_dsl_rule_qed": 74,
   "sel_dsl_rules": 74,
   "sel_rules_simplified_basis": 74,
-  "selector_lines_code": 17912,
-  "selector_lines_total": 28533,
+  "selector_lines_code": 17938,
+  "selector_lines_total": 28559,
   "selector_wildcard_arms_code": 55,
   "selector_wildcard_arms_total": 90,
   "version": "0.57.0",
